Seperate Faster Whisper and Openai Whisper with env variable
